extracted Superblock reading/writing into separate file
